Redesign `show note` page (#76)

+++ app/Http/Controllers/NoteController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Note;
+use App\Services\NotesRepository;
 use Illuminate\Contracts\View\Factory as ViewFactory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Support\Facades\Crypt;
@@ -50,53 +51,63 @@ class NoteController extends Controller
     }
 
     /**
-     * Display the specified resource.
+     * Показываем предупреждение перед расшифровкой заметки (или сообщение,
+     * что заметка не существует).
      *
-     * @param $slug
+     * @param string $slug
+     * @param NotesRepository $notes_repository
      * @return ViewFactory|View
      */
-    public function openLink($slug): View|ViewFactory
-    {
-        $note = Note::where('slug', $slug)->first();
+    public function openLink(
+        string $slug,
+        NotesRepository $notes_repository
+    ): View|ViewFactory {
+        $note = $notes_repository->findBySlug($slug);
 
-        if ($note->expiration_date ?? '' && $note->expiration_date < now()) {
+        if ($note instanceof Note && $note->expiration_date !== null && $note->expiration_date < now()) {
             $note->delete();
-
-            return view('password-note');
+            $note = null;
         }
 
-        $note->password ?? '' === 'none' ? $password = false : $password = true;
-
-        return view('password-note', compact('note', 'password'));
+        return view('note.before_watching', [
+            'hide_footer' => true,
+            'note' => $note,
+        ]);
     }
 
+    /**
+     * @param $slug
+     */
     public function decrypt($slug)
     {
+        // TODO to request
         request()->validate([
             'decrypt_password' => 'string|max:100',
-        ], [
-            'decrypt_password.string' => 'Le champ est vide ou une erreur est survenue',
-            'decrypt_password.max' => 'Le mot de passe ne peut pas faire plus de 100 caractères',
         ]);
 
+        // TODO Use repository
         $note = Note::where('slug', $slug)->firstOr(
             function () {
-                return back()->withErrors(['404' => 'Désolé cette note n\'existe pas ou elle a déjà été lue']);
+                return back()->withErrors(['404' => 'Заметка не существует, уже прочитана или срок ее действия истек']);
             }
         );
 
-        if ($note->password !== 'none') {
+        // TODO Simplify this block
+        if ($note->password !== null) {
             if (Hash::check(request()->decrypt_password, $note->password)) {
-                $note->text = Crypt::decryptString($note->text);
+                $note_text = Crypt::decryptString($note->text);
             } else {
-                return back()->withErrors(['bad_password' => 'Mot de passe incorrect']);
+                return back()->withErrors(['bad_password' => 'Password incorrect']);
             }
         } else {
-            $note->text = Crypt::decryptString($note->text);
+            $note_text = Crypt::decryptString($note->text);
         }
 
         $note->delete();
 
-        return view('show-note', compact('note'));
+        return view('note.show', [
+            'hide_footer' => true,
+            'note_text' => $note_text ?? null,
+        ]);
     }
 }
